Build: use the list of no-build targets to set IS_ANYTHING_TO_BUILD
Currently, we are supposed to build binaries via limited numbers
of pre-defined phony targets.

For example,

  make CROSS_COMPILE=aarch64-linux-gnu- bl1

works, but

  make CROSS_COMPILE=aarch64-linux-gnu- build/fvp/release/bl1.bin

fails because this does not set IS_ANYTHING_TO_BUILD flag.

This restriction keeps platforms from generating plat-specific
images based on the standard images.

Having such phony targets as "bl1", "bl2" in the dependency is not
good; any target depending on phony targets is always rebuilt.

Flip the logic and specify the no-build targets.  Also, this will
be more stable because we rarely add no-build targets.

While we are here, delete the unneeded match_goals helper because
$(filter ...) and $(filter-out ...) are clever enough to work
without $(foreach ...) loop.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>